+++ services/logme/CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.29.0
+- `v1api`
+  - **Improvement**: Use new `WaiterHandler` struct in the LogMe WaitHandler    
+  - **Breaking change:** Change return type of `wait.DeleteInstanceWaitHandler()` to `*wait.AsyncActionHandler[logme.Instance]`
+
 ## v0.28.2
 - **Dependencies:** Bump STACKIT SDK core module from `v0.25.0` to `v0.26.0`